dashboard_backend/requirements.txt
-r ../ba-ner/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
